modulate code using data-grab, dom-utils and listeners
